Changed BallActor and BallMove to BobberActor and BobberMove. Recolored the ball to look like a bobber.
