+++ Game.cpp
@@ -18,7 +18,7 @@
 #include "FPSActor.h"
 #include "PlaneActor.h"
 #include "TargetActor.h"
-#include "BallActor.h"
+#include "BobberActor.h"
 #include "PauseMenu.h"
 #include <SDL/SDL.h>
 #include <SDL/SDL_ttf.h>
