+++ Game.vcxproj.filters
@@ -70,10 +70,10 @@
     <ClCompile Include="FPSCamera.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="BallActor.cpp">
+    <ClCompile Include="BobberActor.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="BallMove.cpp">
+    <ClCompile Include="BobberMove.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="BoxComponent.cpp">
@@ -165,10 +165,10 @@
     <ClInclude Include="FPSCamera.h">
       <Filter>Source Files</Filter>
     </ClInclude>
-    <ClInclude Include="BallActor.h">
+    <ClInclude Include="BobberActor.h">
       <Filter>Source Files</Filter>
     </ClInclude>
-    <ClInclude Include="BallMove.h">
+    <ClInclude Include="BobberMove.h">
       <Filter>Source Files</Filter>
     </ClInclude>
     <ClInclude Include="BoxComponent.h">
